v1.5
Changes:
FULL REDESIGNED THE APPLICATION.
FULL UPDATE FUNCTIONAL THE APPLICATION.
Class "DialogDFD" and dfd format - replace arrays "request" int on the arrays "request" string AND replace "nextIdElement" int on the "nextElement" string.
Visualization code in the project.
Replace "root" to the "TypeUser".
Reduced methods open and create file.dfd.
Added:
Method "ToLanguage" in the class "FileManagerDLAG"
Method "ToTypeUser" in the class "Roots"
Methods "UpdateWindowElements" and "SaveFileBeforeClosing" in the "VisualBindings" class.
New button in the top menu "Close" for closing the file.dfd in the application.

And much more...